🗃Refactor: Api function 분리
