+++ src/common/CommonUtils.ts
