+++ src/common/index.ts
@@ -1,1 +1,2 @@
-export * from '@/common/utils/utils';
+export * from '@/common/CommonUtils';
+export * from '@/common/ApiUtils';
